Revert "Feature/us 5.1 refactor"
