--- biblioteca/src/main/java/com/normalizadas/controller/BooksController.java
@@ -1,25 +1,0 @@
-package com.normalizadas.controller;
-
-import java.util.List;
-
-import com.normalizadas.model.Book;
-import com.normalizadas.model.BookDAOInterface;
-
-public class BooksController {
-
-    private BookDAOInterface bookDAOInterface;
-
-    public BooksController(BookDAOInterface bookDAOInterface){
-        this.bookDAOInterface = bookDAOInterface;
-    }
-
-    public List<Book> getBooksbyGenres(String genre){
-        List<Book> books = bookDAOInterface.getBooksbyGenres(genre);
-        return books;
-    }
-
-    public List<Book> getBooksbyAuthors(String author){
-        List<Book> books = bookDAOInterface.getBooksbyAuthors(author);
-        return books;
-    }
-}
