+++ biblioteca/src/main/java/com/normalizadas/dbConnection.java
@@ -5,8 +5,7 @@ import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
-
-import com.normalizadas.config.DBManager;
+import com.normalizadas.dbConnection;
 
 
 public class dbConnection {
