--- biblioteca/src/main/java/com/normalizadas/model/Author.java
@@ -1,30 +1,0 @@
-package com.normalizadas.model;
-
-public class Author {
-    private int id;
-    private String name;
-
-    public Author(){}
-
-    public Author( String name) {
-        this.name = name;
-    }
-
-
-    public int getId() {
-        return this.id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return this.name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-}
