--- biblioteca/src/main/java/com/normalizadas/model/AuthorDAOInterface.java
@@ -1,9 +1,0 @@
-package com.normalizadas.model;
-
-import java.util.List;
-
-public interface AuthorDAOInterface {
-
-    List<Author> getAuthors(int id);
-    
-}
