--- biblioteca/src/main/java/com/normalizadas/model/BookDAOInterface.java
@@ -1,8 +1,0 @@
-package com.normalizadas.model;
-
-import java.util.List;
-
-public interface BookDAOInterface {
-    List<Book> getBooksbyGenres(String genre);
-    List<Book> getBooksbyAuthors(String author);
-}
