--- biblioteca/src/main/java/com/normalizadas/model/GenreDAOInterface.java
@@ -1,8 +1,0 @@
-package com.normalizadas.model;
-
-import java.util.List;
-
-public interface GenreDAOInterface {
-
-    List<Genre> getGenres(int id);
-}
